ui changed

+++ frontend/src/components/Admin/ProductsTable/Delete.js
@@ -1,18 +1,32 @@
 import React from "react";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { Button, Icon } from "@mui/material";
+import { Icon } from "@mui/material";
 import { ProductsApi } from "../../../api/Products";
+import Swal from "sweetalert2";
+import "sweetalert2/src/sweetalert2.scss";
 const Delete = ({ id, onFinish }) => {
    const deleteHandler = async () => {
-     await ProductsApi.delete(id);
-     onFinish()
-      //let deleted = Object.values(products).filter((el) => el.id !== id);
-      // setProducts(deleted);
+      Swal.fire({
+         title: "مطمئنی؟",
+         text: "نمیتونی برش گردونی هااا!!!",
+         icon: "warning",
+         showCancelButton: true,
+         cancelButtonText: "نه پشیمون شدم",
+         confirmButtonColor: "#6d23b6",
+         cancelButtonColor: " #ff5d8f",
+         confirmButtonText: "آره، پاکش کن",
+      }).then(async(result) => {
+         if (result.isConfirmed) {
+            Swal.fire("پاک شد", "success");
+            await ProductsApi.delete(id);
+            onFinish();
+         }
+      });
    };
    return (
       <div>
          <Icon onClick={() => deleteHandler(id)}>
-            <DeleteIcon />
+            <DeleteIcon sx={{ color: "#6d23b6" }} />
          </Icon>
       </div>
    );
